rename code_model_api -> code_model
